Move Hotbar types to seperate file
